% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Clustering-class}
\alias{Clustering-class}
\alias{Clustering}
\title{Class Clustering}
\description{
Partition data into clusters
}
\details{
Object of class \code{Clustering} containing clustering solution of cumulative dataset D_i. The object of class
\code{Clustering} comprise of four slots. Slot \code{Clusters} contain data items of each cluster, slot \code{Centers} contain
cluster centers, slot \code{k} contain the number of centers, while slot \code{clusterMem} contain cluster memberships vector.
}
\section{Slots}{

\describe{
\item{\code{Cluster}}{List of matrices, where each element of the list include data items belonging to the corresponding cluster.}

\item{\code{Centers}}{Matrix of cluster centers.}

\item{\code{k}}{Number of centers.}

\item{\code{clusterMem}}{Numeric vector of cluster membership.}
}}

