% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/Methods.R
\docType{methods}
\name{Clusters}
\alias{Clusters}
\alias{Clusters,Clustering,matrix,numeric-method}
\alias{Clusters,Clustering-method}
\title{Clustering.}
\usage{
Clusters(object, x, k)

\S4method{Clusters}{Clustering,matrix,numeric}(object, x, k)
}
\arguments{
\item{object}{An object of class \code{Clustering}.}

\item{x}{Numeric matrix of data.}

\item{k}{Number of centers.}
}
\value{
An object of class \code{Clustering}
}
\description{
Initialize slots of class \code{Clustering} by partitioning the dataset into \code{k} clusters.
}
\details{
Runs \code{cclust} function from "flexclust" package with default settings i.e.
method = "kmeans", dist = "euclidean", and partition the dataset. Returns object of class \code{Clustering}.
}
