% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2D.R
\docType{data}
\name{Data2D}
\alias{Data2D}
\title{Synthetic Datasets (Two Dimensional)}
\format{
A data frame
\describe{
\item{x1}{X1.}
\item{x2}{X2.}
\item{class}{Class membership.}
}
}
\usage{
Data2D
}
\description{
A list of datasets generated at four time points containing two variables and cluster membership at each point.
}
\keyword{datasets}
