% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data3D.R
\docType{data}
\name{Data3D}
\alias{Data3D}
\title{Synthetic Datasets (Three Dimensional)}
\format{
A data frame
\describe{
\item{x1}{X1.}
\item{x2}{X2.}
\item{x3}{X3.}
\item{class}{Class membership.}
}
}
\usage{
Data3D
}
\description{
A list of datasets generated at four points containing three variables and cluster membership at each point.
}
\keyword{datasets}
