% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/Methods.R
\docType{methods}
\name{extTransitionCan}
\alias{extTransitionCan}
\alias{extTransitionCan,TransitionCan-method}
\alias{extTransitionCan-method}
\title{External Transition Candidate.}
\usage{
extTransitionCan(object)

\S4method{extTransitionCan}{TransitionCan}(object)
}
\arguments{
\item{object}{An object of class Transitioncan}
}
\value{
Return an object of class TransitionCan
}
\description{
This S4 method trace cluster solutions of dynamic dataset, and identify the candidates
that experience external transition from first clustering and emerged at second clustering.
}
