% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/Methods.R
\docType{methods}
\name{extTransitionCount}
\alias{extTransitionCount}
\alias{extTransitionCount,TransitionCount-method}
\alias{extTransitionCount-method}
\title{External Transition Count}
\usage{
extTransitionCount(object)

\S4method{extTransitionCount}{TransitionCount}(object)
}
\arguments{
\item{object}{An object of class \code{Transitioncount}}
}
\value{
Return an object of class \code{TransitionCount}
}
\description{
Trace cluster solutions of dynamic datasets and count the number of clusters that experiences
external transition from first clustering. The external transition includes survived, split into various daughters,
spliced into one, disappeared, and newly emerged candidates.
}
