% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/Methods.R
\docType{methods}
\name{internalTransition}
\alias{internalTransition}
\alias{internalTransition,intTransitionCan-method}
\alias{internalTransition-method}
\title{Internal Transition Candidates.}
\usage{
internalTransition(object)

\S4method{internalTransition}{intTransitionCan}(object)
}
\arguments{
\item{object}{An object of class \code{intTransitionCan}}
}
\value{
Return an object of class \code{intTransitionCan}
}
\description{
This method identify internal transition of the survived clusters,
obtained from 'extTransitionCan()' method.

Trace clustering solutions of cumulative datasets and identify the survived clusters
experiencing Internal transitions. Internal transition includes the change in location and density
of the survived candidates.
}
