% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{intTransitionCan-class}
\alias{intTransitionCan-class}
\alias{intTransitionCan}
\title{Internal Transition Candidates}
\arguments{
\item{object}{An object of class Transitioncan}
}
\description{
Class containing results of Internal Transition of survived clusters from first clustering \eqn{\xi_1}.
}
\section{Slots}{

\describe{
\item{\code{Location.diff}}{Vector of integers containing difference in location (= Distance bw cluster centers/min(rx,ry)).}

\item{\code{Compactness.diff}}{Vector of integers containing Change in density of survived clusters (d(rx, ry)).}

\item{\code{Location_thrHold}}{Minimum value of threshold for shift in location.}

\item{\code{Density_thrHold}}{Minimum value of threshold for change in density.}

\item{\code{ShiftLocCan}}{Vector of integers containing Survived candidates with shift in their location.}

\item{\code{NoShiftLocCan}}{Vector of integers containing Survived candidates with no Shift in their Location.}

\item{\code{MoreCompactCan}}{Vector of integers containing Survived Candidates Which becomes more compact.}

\item{\code{MoreDiffuseCan}}{Vector of integers containing Survived Candidates Which becomes more diffuse.}

\item{\code{NoChangeCompactCan}}{Vector of integers containing Survived candidates with no change in compactness.}
}}

