% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Monic}
\alias{Monic}
\title{An S4 class that contain time steps}
\arguments{
\item{object}{An object of class \code{Transitioncan}}
}
\description{
An S4 class that contain time steps
}
\section{Slots}{

\describe{
\item{\code{TimeStep}}{Time Steps}
}}

