% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/Methods.R
\docType{methods}
\name{Overlap}
\alias{Overlap}
\alias{Overlap,OverLap,Clustering,Clustering-method}
\alias{OverLap,}
\alias{Clustering,}
\alias{Clustering-method}
\alias{Overlap,OverLap,ANY,ANY-method}
\alias{OverLap,ANY,ANY-methods}
\title{Overlap}
\usage{
Overlap(object, e1, e2)

\S4method{Overlap}{OverLap,Clustering,Clustering}(object, e1, e2)

\S4method{Overlap}{OverLap,ANY,ANY}(object, e1, e2)
}
\arguments{
\item{object}{An object of class \code{OverLap}}

\item{e1}{An object of class \code{Clustering}, or any object that can be coerced, such as list of matrices
or data frames that contain clusters from first clustering.}

\item{e2}{An object of class \code{Clustering}, or any object that can be coerced, such as list of matrices
or data frames that contain clusters from second clustering.}
}
\value{
Return an object of class \code{OverLap}.
}
\description{
Initialize slots of class \code{OverLap} by importing clustering solutions of dynamic
datasets at two consecutive time points. Clusters at each time point should be provided as a list of
matrices, where each matrix contains dataset belongs to the corresponding cluster.
}
