% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{show,Monic-method}
\alias{show,Monic-method}
\alias{show-method}
\title{Show Method for output}
\usage{
\S4method{show}{Monic}(object)
}
\arguments{
\item{object}{An object of class Monic}
}
\description{
Show Method for output
}
