% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{TransitionCount-class}
\alias{TransitionCount-class}
\alias{TransitionCount}
\title{External Transition Count}
\description{
Trace cluster solutions of dynamic datasets at consecutive
time points and counts the clusters that experiences external transition.
External transition includes Survive, Split, Merge, newly emerged, and Died candidates.
}
\section{Slots}{

\describe{
\item{\code{Survive}}{Number of candidates survive from first clustering \eqn{\xi_1}.}

\item{\code{Split}}{Number of candidates from first clustering \eqn{\xi_1} that split into several daughter clusters at second clustering \eqn{\xi_2}.}

\item{\code{Merge}}{Number of candidates from first clustering \eqn{\xi_1} that merge toghter at second clustering \eqn{\xi_2}.}

\item{\code{Died}}{Number of candidates from first clusterin \eqn{\xi_1} that disapeared at second clustering \eqn{\xi_2}.}

\item{\code{SurvivalRatio}}{Ratio of survive clusters to total number of clusters from first clusering \eqn{\xi_1}.}

\item{\code{AbsorptionRatio}}{Ratio of Merged clusters to total number of clusters from first clusering \eqn{\xi_1}.}

\item{\code{passforwardRatio}}{Sum of SurvivalRatio and AbsorptionRatio.}

\item{\code{Survival_thrHold}}{Threshold for survival of clusters.}

\item{\code{Split_thrHold}}{Threhold for split of clusters.}

\item{\code{Cluster_Tracex}}{Vector containing each cluster result from first clustering \eqn{\xi_1}.}
}}

